Major changes to MainScreen and ExtensionsScreen.kt ExtensionsScreen has been merged with MainScreen, in the sense that MainScreen pulls Extensions from ExtensionsScreen.kt and adds them to it's own Column Links has been renamed to AppInfoChip
